Changes to common files for 441e.

+++ src/common/net-unix.c
@@ -574,7 +574,7 @@ int	port;
     peer.sin_family = AF_INET;
     peer.sin_port   = htons(port);
     peer.sin_addr.s_addr = inet_addr(host);
-    if (peer.sin_addr.s_addr == (unsigned int)-1)
+    if (peer.sin_addr.s_addr == (int)-1)
     {
 	hp = gethostbyname(host);
 	if (hp == NULL)
@@ -1714,7 +1714,7 @@ int	port;
     addr_in.sin_family          = AF_INET;
     addr_in.sin_port            = htons(port);
     /**/addr_in.sin_addr.s_addr 	= inet_addr(host);
-    if (addr_in.sin_addr.s_addr == (unsigned int)-1)
+    if (addr_in.sin_addr.s_addr == (unsigned long)-1)
     {
 #ifdef SERVER 
 	printf("DgramConnect called with hostname %s.\n", host);
@@ -1816,7 +1816,7 @@ char	*host, *sbuf;
     else
     {
 	the_addr.sin_addr.s_addr 	= inet_addr(host);
-	if (the_addr.sin_addr.s_addr == (unsigned int)-1)
+	if (the_addr.sin_addr.s_addr == (int)-1)
 	{
 #ifdef SERVER
 	    printf("DgramSend called with host %s\n", host);
@@ -1949,7 +1949,7 @@ char	*rbuf;
     struct sockaddr_in	tmp_addr;
 
     tmp_addr.sin_addr.s_addr = inet_addr(from);
-    if (tmp_addr.sin_addr.s_addr == (unsigned int)-1)
+    if (tmp_addr.sin_addr.s_addr == (int)-1)
     {
 #ifdef SERVER
 	printf("DgramReceive called with host %s.\n", from);
